Fixes for half/byte stores
RISCV expects the data to be copied across the rest of the word
for half/byte stores. So, if it is SB then the 8 least significant
bits of what should be stored must be copied across the rest of the
word.
byte_en must be flipped if storing. This is because the byte_en
is applied to loads BEFORE the endian flip, but applied to stores
AFTER the endian flip