Add necessary files for sparsity optimizations
Functionality is not included, but added all the necessary modules, interfaces,
and scripts that integrate the sparsity optimizations with the pipeline. the
SPARCE_ENABLED parameter can be set to "enabled" or "disabled", and this chooses between
the sparce_enabled and sparce_disabled modules in the sparce_wrapper module.

additionally, created the sparce_internal_if interface that contains all the modports for
the blocks in the SparCE paper