Connect signals for sparsity optimizations to proper place in pipeline.
The sparce_internal_if was connected to the proper signals from sparce_pipeline_if.
Additionally, the pc in the pipeline was muxed to include the program flow changes
from the sparsity optimizations. Finally, the input signals to the sparsity
optimizations were connected to signals in the tspp_execute_stage and
tspp_fetch stage.