Add SpRF testbench, along with supporting modifications to other files
The testbench was created with several tasks specified in the Test Plan
on Confluence. The wscript in the source_code folder was modified to
contain build information for the sprf file and testbench. Finally,
added support for the sprf testbench in the run_tests script by adding
it to the list of supported sparsity tests.