The 'sparce_skip_range.S' is completed and tested to work correctly.
The test case loops through all possible skip values between 0 and 31.
The sasa_table is configured to be a direct mapped configuration
because there is an issue with the replacement policy for the 4-way
associative configuration. The 5th entry seemed to not be written
correctly when the table for that index is occupied.