Fix bug in SASA table involving duplicate writes, and update tb
If the SASA table had multiple writes to the same tag/index pair,
the table would insert the new writes into the oldest entry via
the LRU replacement policy, even if this would result in duplicate
tag/index pairs. This update checks to see if there is an existing
entry in the table, and has that entry replaced, instead. An
additional test case was added to the SASA table testbench that
verifies that duplicate writes occur as expected