Fix for RV32C incorrect execution.
Occasionally, when RV32C is able to provide an instruction early
(i.e. buffered compressed instruction), execution could be skipped
due to some validity flags relying on waiting for I-fetch. This provides
a small fix to the hazard unit that detects the "early finish" condition
and allows the updates the same as if instruction fetch had completed.