Initial commit of 3-stage pipeline: All tests passing
This adds a 3-stage pipeline, defined as follows:

 Fetch | Decode/Execute | Memory/Writeback

There is a forwarding path from the second latch into the decode/execute
stage. However, memory references cannot be forwarded due to critical
path issues, so load->use and csr->use hazards will incur 1 extra cycle
of stalling (beyond usual delay of load) for dependent instructions.

Some rough estimates of frequency (standalone) are:
- 70MHz max on FPGA
- 200MHz+ feasible in ASIC
This implies that 50MHz FPGA speed and 100MHz ASIC speed should be attainable,
with more room to push the latter depending on the rest of the system.

All the RV32I tests pass. Remaining TODOs are:
- Alter the TB to allow testing of forwarding. Currently, memory references take
so long that there are never back-to-back instructions in the pipeline. While
this is somewhat realistic (AFT will have these delays), the addition of a cache
(and prefetching effect of RV32C) will allow single-cycle instruction hits.
- Implement exceptions/interrupts. This is currently untested.
- Allow RV32C. Theoretically, this shouldn't cause any problems, since
the parts of the pipeline that touch the decompressor were not altered.
- RV32M. There needs to be a decision about whether RISC-MGMT should be
supported for this pipeline. Ideally it should be added, but this is also
a chance to remove the standard extensions from RISC-MGMT, and change it to
be only for custom instructions if that is desired.