stage3: Interrupt test passing
This changes the logic for stalling/flushing so that on receiving an
interrupt, the currently-executing memory instruction is allowed to finish
(if such an instruction exists), then the oldest PC in the pipeline is taken.

Since this is asynchronous, there is no guarantee the M-stage has a valid PC,
so priority logic is needed to select from the 3 PCs in the pipeline. This is
simpler than latching the next PC of the last valid instruction since it doesn't
require instruction-specific knowledge (e.g. control flow target, compressed, etc.).

Signals were added to the pipeline to track when an instruction is valid as well, since
insn == 0 cannot be assumed to be a pipeline bubble instead of a nop instruction.