standard_core: Add test version without memory controller
This adds a second version of RISCVBusiness for testing the core that does not
have a memory controller. This allows direct access to the buses for testing with
different latencies. Currently, it just does immediate latencies to mimic having
perfect caches.

Additional work would be to add the ability to simulate hits/misses by having
a random chance of getting bad latency, independently for I/D streams.