stage3: Fix bugs resulting from back-to-back instruction execution
This fixes bugs related to hazards occurring only in the case of back-to-back
execution of instructions. The fixes were:
- Fixing forwarding unit assignments to allow detection of forwarding conditions correctly
- Forcing ifence to flush the pipeline and re-fetch in-flight instructions that may
no longer be valid
- Fix RV32C to obey the pipeline control signals. Previously, it ignored the "pc_en" signal,
which led to cases where instructions would be skipped if the first pipeline latch was stalled
while RV32C wanted to advance.

Remaining items to test/implement:
- Flushing for selected CSR writes. Need a list of such instructions, but at minimum this should
include the PMP/PMA configuration registers.
- Testing variable latencies, instead of only fixed slow/fast latencies