stage3: RV32M integration
This commit adds RV32M to the stage3 pipeline. This does not use
RISC-MGMT, instead opting for a wrapper with enable/disable like
RV32C. In discussions with the team, this seems more manageable than
trying to fit more complex extensions (that include state) into RISC-MGMT.

In the future, RISC-MGMT should be integrated to allow custom instructions.

All tests for RV32IMC pass.